pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "email-assistant"
version = "0.1.0"
description = "AI-Powered Email Assistant - Multi-agent LangGraph pipeline"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-cohere>=0.2.0",
    "streamlit>=1.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["email_assistant*"]
